requirements.txt
google-genai>=0.2.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
pytest>=7.4.0

